+++ eventApp/app/Http/Controllers/InvoiceController.php
@@ -8,7 +8,9 @@ use App\Http\Requests\StoreInvoiceRequest;
 use App\Http\Requests\UpdateInvoiceRequest;
 use App\Http\Resources\InvoiceResource;
 use App\Http\Resources\InvoiceCollection;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Hash;
 
 class InvoiceController extends Controller
 {
@@ -28,7 +30,7 @@ class InvoiceController extends Controller
         }
         if (!$member) {
             return response()->json([
-                'message' => 'No events found for the user',
+                'message' => 'No invoices found for the user',
                 'data' => []
             ], 404);
         }
@@ -58,16 +60,57 @@ class InvoiceController extends Controller
      */
     public function store(StoreInvoiceRequest $request)
     {
-        return new InvoiceResource(Invoice::create($request->all()));
+        $user = auth()->user();
+        $userid = auth()->id();
 
+        $isAdmin = DB::table('members')->where('id', $userid)->first();
+        if ($isAdmin->role === 'A') {
+            $request->validated();
+            $invoice = Invoice::create([
+                'status' => $request['status'],
+                'memberID' => $request['memberID'],
+                'billedDate' => $request['billedDate'],
+                'paidDate' => $request['paidDate'],
+            ]);
+            if($invoice){
+                return response()->json([
+                    'message' => 'Successfully created invoice!',
+                    'user' => $invoice
+                ]);
+            }else{
+                return response()->json([
+                    'message' => 'Fail created invoice!'
+                ]);
+            }
+        }else{
+            return response()->json([
+                'message' => 'Not allowed to create invoice!',
+            ]);
+        }
+
+
+
+        //return new InvoiceResource(Invoice::create($request->all()));
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Invoice $invoice)
+    public function show($id)
     {
-        return new InvoiceResource($invoice);
+        $user = auth()->user();
+        $userId = auth()->id();
+        $isAdmin = DB::table('members')->where('id', $userId)->first();
+        $invoice = DB::table('invoices')->where('id', $id)->first();
+        if($isAdmin->role === 'A'){
+            return response()->json(['member' => $invoice], 200);
+
+        }else{
+            return response()->json([
+                'message' => 'You are not allowed to view this invoice!',
+            ]);
+        }
+        //return new InvoiceResource($invoice);
     }
 
     /**
@@ -83,7 +126,21 @@ class InvoiceController extends Controller
      */
     public function update(UpdateInvoiceRequest $request, Invoice $invoice)
     {
-        $invoice->update($request->all());
+        $user = Auth::user();
+        $userid = Auth::id();
+        $data = $request->all();
+        $isAdmin = DB::table('members')->where('id', $userid)->first();
+        $memberID = $isAdmin->id;
+        if (!$memberID) {
+            return response()->json(['message' => 'Invoice not found'], 404);
+        }
+        if ($isAdmin->role === 'A') {
+            $invoice->update($request->all());
+            return response()->json(['message' => 'Invoice updated successfully']);
+        }else{
+            return response()->json(['message' => 'You are not allowed to edit this invoice!'], 404);
+        }
+        //$invoice->update($request->all());
 
     }
 
@@ -92,16 +149,25 @@ class InvoiceController extends Controller
      */
     public function destroy(Invoice $invoice)
     {
-        try {
-            $invoice->delete();
-            return response()->json([
-                'message' => 'Invoice deleted successfully!'
-            ], 200);
-        } catch (\Exception $e) {
-            return response()->json([
-                'error' => 'Something went wrong!',
-                'message' => $e->getMessage()
-            ], 500);
+        $user = auth()->user();
+        $userid = auth()->id();
+        $isAdmin = DB::table('members')->where('id', $userid)->first();
+        if($isAdmin->role === 'A'){
+            try {
+                $invoice->delete();
+                return response()->json([
+                    'message' => 'Invoice deleted successfully!'
+                ], 200);
+            } catch (\Exception $e) {
+                return response()->json([
+                    'error' => 'Something went wrong!',
+                    'message' => $e->getMessage()
+                ], 500);
+            }
+        }else{
+            return response()->json(['message' => 'You are not allowed to delete this invoice!'], 404);
+
         }
+
     }
 }
